requirements.txt
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
feedparser==6.0.12
filelock==3.20.0
filetype==1.2.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
Js2Py_3.13==0.74.1
llvmlite==0.45.1
lxml==6.0.2
m3u8==6.0.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6
numba==0.62.1
numpy==2.3.5
openai==2.9.0
openai-whisper==20250625
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyjsparser==2.7.1
pyotp==2.9.0
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
twikit==2.3.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.0
webvtt-py==0.5.1
wheel==0.45.1
yt-dlp==2025.11.12
